added helper function files and update folder structure
