+++ src/App.js
@@ -3,12 +3,13 @@ import ReactGA from 'react-ga';
 import './App.css';
 import NavBar from './Components/Navbar/Navbar';
 import { Box, Paper, InputLabel, MenuItem, FormControl, Select, IconButton, Menu, Typography, Grid, makeStyles } from '@material-ui/core';
-import StickyTable from './Components/FixedHeaderTable';
-import LineChart from './Components/LineChart';
-import LineChartTwo from './Components/LineCartTwo';
-import StackedBarChar from './Components/StackedBarChart';
-import MixedBarChart from './Components/MixedBarChart';
+import StickyTable from './Components/Table/FixedHeaderTable';
+import LineChart from './Components/Line Charts/LineChart';
+import LineChartTwo from './Components/Line Charts/LineCartTwo';
+import StackedBarChar from './Components/Bar Charts/StackedBarChart';
+import MixedBarChart from './Components/Bar Charts/MixedBarChart';
 import DropbdownButton from './Components/Dropdown Button/DropDownButton';
+import { calculateWorldwideTotalCases, calculateWorldwideTotalDeaths } from './libs/helpers';
 const GA_ID = `${process.env.REACT_APP_GA_KEY}`;
 const HOST_KEY = `${process.env.REACT_APP_HOST}`;
 const API_KEY = `${process.env.REACT_APP_COVID19_API_KEY}`;
@@ -16,33 +17,6 @@ const API_KEY = `${process.env.REACT_APP_COVID19_API_KEY}`;
 const initializeAnalytics = () => {
   ReactGA.initialize(GA_ID)
   ReactGA.pageview('/')
-}
-
-const calculateWorldwideTotalCases = obj => {
-  if (obj === undefined) {
-    return null
-  } else {
-    const arrayObject = obj.map(result => {
-      return result.cases.total
-    })
-    // console.log(arrayObject)
-    const total = arrayObject.reduce((a, b) => a + b / 2)
-    const roundedTotal = Math.round(total).toLocaleString()
-    return roundedTotal
-  }
-}
-
-const calculateWorldwideTotalDeaths = obj => {
-  if (obj === undefined) {
-    return null
-  } else {
-    const arrayObject = obj.map(result => {
-      return result.deaths.total
-    })
-    const deathTotal = arrayObject.reduce((a, b) => a + b / 2)
-    const roundedDeathTotal = Math.round(deathTotal).toLocaleString()
-    return roundedDeathTotal
-  }
 }
 
 function App() {
@@ -59,10 +33,8 @@ function App() {
 
   const updateStats = event => {
     const value = event.target.value
-
     setCountry(value);
     setQuery(value);
-
     ReactGA.event({
       category: 'Country Selected',
       action: 'Button Click',
@@ -119,20 +91,18 @@ function App() {
     fetchData()
   }, [query]);
 
-
-
   return (
     <div className="App">
       <NavBar />
         <Grid item xs={12} container direction="row" justify="flex-start" alignItems="center" style={{paddingLeft: 50, paddingRight: 50, marginTop: 10}}>
           <DropbdownButton handleChange={updateStats} country={country} />
-          <Typography style={{ marginLeft: 20 }}>Get statistics for all countries about COVID-19</Typography>
-          <div style={{ marginLeft: 20 }}> | </div>
-          <Typography style={{ marginLeft: 20 }}>Worldwide Cases: <strong>{totalCases}</strong></Typography>
-          <div style={{ marginLeft: 20 }}> | </div>
-          <Typography style={{ marginLeft: 20 }}>Worldwide Deaths: <strong>{totalWorldwideDeaths}</strong></Typography>
+          <Typography style={{ paddingLeft: 20 }}>Get statistics for all countries about COVID-19</Typography>
+          <div style={{ paddingLeft: 20 }}> | </div>
+          <Typography style={{ paddingLeft: 20 }}>Worldwide Cases: <strong>{totalCases}</strong></Typography>
+          <div style={{ paddingLeft: 20 }}> | </div>
+          <Typography style={{ paddingLeft: 20 }}>Worldwide Deaths: <strong>{totalWorldwideDeaths}</strong></Typography>
           {/* <PieChart data={countryHistory} /> */}
-            <Grid item xs={12} container direction="row" justify="space-evenly">
+            <Grid item xs={12} container direction="row" justify="space-evenly" style={{marginBottom: 10}}>
               <Paper>
                 <StackedBarChar data={countryHistory} />
               </Paper>
