--- src/Components/StackedBarChart.js
@@ -1,68 +1,0 @@
-import React from 'react';
-import {
-    BarChart, Bar, Cell, XAxis, YAxis, CartesianGrid, Tooltip, Legend,
-} from 'recharts';
-
-const data = [
-    {
-        name: 'Page A', uv: 4000, pv: 2400, amt: 2400,
-    },
-    {
-        name: 'Page B', uv: 3000, pv: 1398, amt: 2210,
-    },
-    {
-        name: 'Page C', uv: 2000, pv: 9800, amt: 2290,
-    },
-    {
-        name: 'Page D', uv: 2780, pv: 3908, amt: 2000,
-    },
-    {
-        name: 'Page E', uv: 1890, pv: 4800, amt: 2181,
-    },
-    {
-        name: 'Page F', uv: 2390, pv: 3800, amt: 2500,
-    },
-    {
-        name: 'Page G', uv: 3490, pv: 4300, amt: 2100,
-    },
-];
-
-const convertObject = obj => {
-    if (obj === undefined) {
-        return null
-    } else {
-        const objectArray = Object.entries(obj).reverse()
-        const arrayObject = objectArray.map(([key, value]) => {
-            return value
-        })
-        return arrayObject
-    }
-}
-
-export default function BarGraph(props) {
-    const data = props.data
-    const sortedData = convertObject(data)
-
-    return (
-        <React.Fragment>
-            <h4 style={{ paddingTop: 10 }}>Cases vs. Deaths</h4>
-            <BarChart
-                width={500}
-                height={300}
-                data={sortedData}
-                margin={{
-                    top: 20, right: 60, left: 20, bottom: 5,
-                }}
-            >
-                <CartesianGrid strokeDasharray="3 3" />
-                <XAxis dataKey="day" />
-                <YAxis />
-                <Tooltip />
-                <Legend />
-                <Bar dataKey="cases.total" stackId="a" fill="#8884d8" />
-                <Bar dataKey="deaths.total" stackId="a" fill="#ffc658" />
-            </BarChart>
-        </React.Fragment>
-    );
-}
-
